+++ src/main/java/com/atlauncher/managers/ModrinthModpackUpdateManager.java
@@ -18,19 +18,66 @@
 package com.atlauncher.managers;
 
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
-import com.atlauncher.App;
-import com.atlauncher.Data;
 import com.atlauncher.data.Instance;
 import com.atlauncher.data.modrinth.ModrinthVersion;
 import com.atlauncher.utils.ModrinthApi;
 
+import io.reactivex.rxjava3.core.Observable;
+import io.reactivex.rxjava3.subjects.BehaviorSubject;
+
 public class ModrinthModpackUpdateManager {
-    public static ModrinthVersion getLatestVersion(Instance instance) {
-        return Data.MODRINTH_INSTANCE_LATEST_VERSION.get(instance);
+    /**
+     * Modrinth instance update checking
+     */
+    private static final Map<UUID, BehaviorSubject<Optional<ModrinthVersion>>>
+        MODRINTH_INSTANCE_LATEST_VERSION = new HashMap<>();
+
+    /**
+     * Get the update behavior subject for a given instance.
+     *
+     * @param instance Instance to get behavior subject for
+     * @return behavior subject for said instance updates.
+     */
+    private static BehaviorSubject<Optional<ModrinthVersion>> getSubject(Instance instance){
+        MODRINTH_INSTANCE_LATEST_VERSION.putIfAbsent(
+            instance.getUUID(),
+            BehaviorSubject.createDefault(Optional.empty())
+        );
+        return MODRINTH_INSTANCE_LATEST_VERSION.get(instance.getUUID());
     }
 
+    /**
+     * Get an observable for an instances update.
+     * <p>
+     * Please do not cast to a behavior subject.
+     * @param instance Instance to get an observable for
+     * @return Update observable
+     */
+    public static Observable<Optional<ModrinthVersion>> getObservable(Instance instance) {
+        return getSubject(instance);
+    }
+
+
+    /**
+     * Get the latest version of an instance
+     * @param instance Instance to get version of
+     * @return Latest version, or null if no newer version is found
+     */
+    public static ModrinthVersion getLatestVersion(Instance instance) {
+        return getSubject(instance).getValue().orElse(null);
+    }
+
+    /**
+     * Check for new updates.
+     * <p>
+     * Updates observables.
+     */
     public static void checkForUpdates() {
         if (ConfigManager.getConfigItem("platforms.modrinth.modpacksEnabled", true) == false) {
             return;
@@ -39,45 +86,20 @@ public class ModrinthModpackUpdateManager {
         PerformanceManager.start();
         LogManager.info("Checking for updates to Modrinth instances");
 
-        boolean refreshInstancesPanel = Data.INSTANCES.parallelStream()
-                .filter(i -> i.isModrinthPack()).map(i -> {
-                    boolean wasUpdated = false;
-
+        InstanceManager.getInstances().parallelStream()
+                .filter(i -> i.isModrinthPack()).forEach(i -> {
                     List<ModrinthVersion> packVersions = ModrinthApi.getVersions(i.launcher.modrinthProject.id);
 
                     if (packVersions == null) {
-                        return false;
+                        return;
                     }
 
                     ModrinthVersion latestVersion = packVersions.stream()
                             .sorted(Comparator.comparing((ModrinthVersion version) -> version.datePublished).reversed())
                             .findFirst().orElse(null);
 
-                    if (latestVersion == null) {
-                        return false;
-                    }
-
-                    // if there is a change to the latestversion for an instance (but not a first
-                    // time write), then refresh instances panel
-                    if (Data.MODRINTH_INSTANCE_LATEST_VERSION.containsKey(i)
-                            && !Data.MODRINTH_INSTANCE_LATEST_VERSION.get(i).id.equals(latestVersion.id)) {
-                        wasUpdated = true;
-                    }
-
-                    // updated if there is no latest version stored yet but the instance has update
-                    if (!Data.MODRINTH_INSTANCE_LATEST_VERSION.containsKey(i)
-                            && !i.launcher.modrinthVersion.id.equals(latestVersion.id)) {
-                        wasUpdated = true;
-                    }
-
-                    Data.MODRINTH_INSTANCE_LATEST_VERSION.put(i, latestVersion);
-
-                    return wasUpdated;
-                }).anyMatch(b -> b);
-
-        if (refreshInstancesPanel) {
-            App.launcher.reloadInstancesPanel();
-        }
+                    getSubject(i).onNext(Optional.ofNullable(latestVersion));
+                });
 
         PerformanceManager.end();
     }